requirements.txt
absl-py==2.1.0
accelerate==0.21.0
ace_tools==0.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
ale-py==0.8.1
alembic==1.15.2
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
autoattack @ git+https://github.com/fra31/auto-attack.git@a39220048b3c9f2cca9a4d3a54604793c68eca7e
autograd==1.6.2
autoray==0.6.12
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
babel==2.16.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
bitsandbytes==0.41.0
bleach==6.2.0
blinker==1.8.2
blis==1.2.0
box2d-py==2.3.5
cachetools==5.3.3
caer==2.0.8
catalogue==2.0.10
certifi==2024.7.4
cffi==2.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==3.0.0
cmake==4.2.1
colabcode==0.1.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cryptography==46.0.3
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
databricks-sdk==0.73.0
datasets==4.0.0
debugpy==1.8.9
decorator==5.1.1
deepmultilingualpunctuation==1.0.1
defusedxml==0.7.1
diastatic-malt==2.15.2
dill==0.3.8
distro==1.9.0
docker==7.1.0
einops==0.6.1
einops-exts==0.0.4
et_xmlfile==2.0.0
etils==1.11.0
executing==2.1.0
Farama-Notifications==0.0.4
fastapi==0.121.2
fastjsonschema==2.21.0
ffmpy==1.0.0
filelock==3.19.1
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.3.0
future==1.0.0
gast==0.6.0
gdown==5.1.0
geotorch==0.3.0
gitdb==4.0.12
GitPython==3.1.45
glfw==2.8.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-pasta==0.2.0
gradio==3.35.2
gradio_client==0.2.9
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.2.0
grpcio==1.65.1
gunicorn==23.0.0
gurobipy==11.0.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
h11==0.14.0
h5py==3.11.0
hashcat==24.11.10
hf-xet==1.2.0
httpcore==0.17.3
httpx==0.24.0
huey==2.5.4
huggingface-hub==0.36.0
HyperPyYAML==1.2.2
idna==3.7
imageio==2.36.1
imbalanced-learn==0.12.4
imblearn==0.0
importlib_metadata==8.7.0
importlib_resources==6.4.5
inquirerpy==0.3.4
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.20
jaxlib==0.4.20
jedi==0.19.2
Jinja2==3.1.6
jiter==0.6.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-http-over-ws==0.0.8
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kagglehub==0.3.13
keras==2.15.0
kiwisolver==1.4.5
langcodes==3.4.1
language_data==1.2.0
latex2mathml==3.78.1
lazy_loader==0.4
libclang==18.1.1
lime==0.2.0.1
linkify-it-py==2.0.3
lit==18.1.8
-e git+https://github.com/microsoft/LLaVA-Med.git@30697ca50b5c29a8e955c99330b259776aef27b9#egg=llava_med
llvmlite==0.45.1
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.6
markdown-it-py==2.2.0
markdown2==2.5.4
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
mediapipe==0.10.18
mistune==3.0.2
mkl==2021.4.0
ml-dtypes==0.3.2
mlflow==3.6.0
mlflow-skinny==3.6.0
mlflow-tracing==3.6.0
mpmath==1.3.0
mujoco==3.2.5
multidict==6.6.3
multiprocess==0.70.16
murmurhash==1.0.10
mypy==1.8.0
mypy-extensions==1.0.0
namex==0.0.8
narwhals==2.14.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.62.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
openai==1.12.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opt-einsum==3.3.0
optree==0.14.1
optuna==4.3.0
orjson==3.11.5
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pathos==0.3.2
pbr==6.1.0
peft==0.4.0
PennyLane==0.40.0
pennylane-qulacs==0.36.0
PennyLane_Lightning==0.40.0
pexpect==4.9.0
pfzy==0.3.4
pillow==11.3.0
pipdeptree==2.23.4
platformdirs==4.3.6
pox==0.3.5
ppft==1.7.6.9
preshed==3.0.9
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.3.2
protobuf==4.25.8
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pybullet==3.2.6
pybullet_envs_gymnasium==0.5.0
pycparser==2.21
pydantic==1.10.25
pydantic_core==2.41.5
pydub==0.25.1
pygame==2.6.1
pygame-ce==2.1.3
Pygments==2.18.0
pyngrok==7.2.1
PyOpenGL==3.1.7
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==2.0.7
python-multipart==0.0.21
pytorch-triton-rocm==3.2.0
pytz==2024.2
PyYAML==6.0.1
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
robustbench @ git+https://github.com/RobustBench/robustbench.git@78fcc9e48a07a861268f295a777b975f25155964
rpds-py==0.21.0
rsa==4.9.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
rustworkx==0.15.1
safetensors==0.5.3
scikit-image==0.26.0
scikit-learn==1.2.2
scipy==1.13.1
scipy-openblas32==0.3.30.0.8
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.1.99
shellingham==1.5.4
Shimmy==1.3.0
shortuuid==1.0.13
six==1.16.0
slicer==0.0.8
smart-open==7.0.5
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechbrain==1.0.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
srsly==2.4.8
stable_baselines3==2.4.0
stack-data==0.6.3
starlette==0.49.3
stevedore==5.3.0
svgwrite==1.4.3
swig==4.3.0
symengine==0.11.0
sympy==1.13.1
tbb==2021.11.0
tcod==16.2.2
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
terminado==0.18.1
thinc==8.3.4
threadpoolctl==3.5.0
tifffile==2025.12.20
tiktoken==0.12.0
timm==0.9.12
tinycss2==1.4.0
tokenizers==0.15.2
toml==0.10.2
tomlkit==0.13.2
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchdiffeq==0.2.5
torchvision==0.15.2+cu118
tornado==6.4.2
tqdm==4.66.4
traitlets==5.14.3
transformers==4.36.2
triton==2.0.0
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.38.0
waitress==3.0.2
wasabi==0.10.1
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.3
widgetsnbextension==4.0.13
wordsegment==1.3.1
wrapt==1.14.2
xgboost==2.1.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.21.0
